+++ CHANGES.rst
@@ -131,6 +131,7 @@ Development
 - Prepare real "light sleep" (WIP)
 - Attempt to reset WiFi connection if scanning fails
 - Add "mboot", the universal MicroPython bootloader
+- Reconfigure watchdog when connecting the device using MiniNet
 
 
 2019-06-22 0.5.1